own file for model.semester and corresponding methods
