+++ model/course.go
@@ -341,8 +341,3 @@ func (c Course) AdminJson(lhs []LectureHall) []gin.H {
 	}
 	return res
 }
-
-type Semester struct {
-	TeachingTerm string
-	Year         int
-}
