+++ model/user.go
@@ -307,29 +307,11 @@ func (u *User) AdministeredCoursesForSemester(year int, term string, context con
 }
 
 func (u *User) AdministeredCoursesForSemesters(firstSemester Semester, lastSemester Semester, semesters []Semester, context context.Context) []Course {
-	var inRangeOfSemesters func(s Semester, firstSemester Semester, lastSemester Semester, semesters []Semester) bool
-	if semesters == nil {
-		inRangeOfSemesters = func(s Semester, firstSemester Semester, lastSemester Semester, semesters []Semester) bool {
-			return !(s.Year < firstSemester.Year || s.Year > lastSemester.Year ||
-				(s.Year == firstSemester.Year && s.TeachingTerm == "S" && firstSemester.TeachingTerm == "W") ||
-				(s.Year == lastSemester.Year && s.TeachingTerm == "W" && lastSemester.TeachingTerm == "S"))
-		}
-	} else {
-		inRangeOfSemesters = func(s Semester, firstSemester Semester, secondSemester Semester, semesters []Semester) bool {
-			for _, semester := range semesters {
-				if s.Year == semester.Year && s.TeachingTerm == semester.TeachingTerm {
-					return true
-				}
-			}
-			return false
-		}
-	}
-
 	var semester Semester
 	administeredCourses := make([]Course, 0)
 	for _, c := range u.AdministeredCourses {
 		semester = Semester{TeachingTerm: c.TeachingTerm, Year: c.Year}
-		if inRangeOfSemesters(semester, firstSemester, lastSemester, semesters) {
+		if semester.InRangeOfSemesters(firstSemester, lastSemester, semesters) {
 			administeredCourses = append(administeredCourses, c)
 		}
 	}
@@ -347,29 +329,11 @@ func (u *User) CoursesForSemesterWithoutAdministeredCourses(year int, term strin
 }
 
 func (u *User) CoursesForSemestersWithoutAdministeredCourses(firstSemester Semester, lastSemester Semester, semesters []Semester, context context.Context) []Course {
-	var inRangeOfSemesters func(s Semester, firstSemester Semester, lastSemester Semester, semesters []Semester) bool
-	if semesters == nil {
-		inRangeOfSemesters = func(s Semester, firstSemester Semester, lastSemester Semester, semesters []Semester) bool {
-			return !(s.Year < firstSemester.Year || s.Year > lastSemester.Year ||
-				(s.Year == firstSemester.Year && s.TeachingTerm == "S" && firstSemester.TeachingTerm == "W") ||
-				(s.Year == lastSemester.Year && s.TeachingTerm == "W" && lastSemester.TeachingTerm == "S"))
-		}
-	} else {
-		inRangeOfSemesters = func(s Semester, firstSemester Semester, secondSemester Semester, semesters []Semester) bool {
-			for _, semester := range semesters {
-				if s.Year == semester.Year && s.TeachingTerm == semester.TeachingTerm {
-					return true
-				}
-			}
-			return false
-		}
-	}
-
 	var semester Semester
 	courses := make([]Course, 0)
 	for _, c := range u.Courses {
 		semester = Semester{TeachingTerm: c.TeachingTerm, Year: c.Year}
-		if inRangeOfSemesters(semester, firstSemester, lastSemester, semesters) && !u.IsAdminOfCourse(c) {
+		if semester.InRangeOfSemesters(firstSemester, lastSemester, semesters) && !u.IsAdminOfCourse(c) {
 			courses = append(courses, c)
 		}
 	}
